Move diagram module to core.modeling
