+++ gaphor/core/modeling/__init__.py
@@ -1,3 +1,4 @@
 from gaphor.core.modeling.element import DummyEventWatcher, Element, Handler
 from gaphor.core.modeling.elementfactory import ElementFactory
+from gaphor.core.modeling.event import *
 from gaphor.core.modeling.presentation import Presentation
