+++ gaphor/core/modeling/diagram.py
