+++ gaphor/core/modeling/tests/test_diagram.py
