+++ utils/model/gen_uml.py
@@ -35,8 +35,6 @@ from gaphor.core.modeling.properties import (
     redefine,
 )
 from gaphor.core.modeling.collection import collection
-from gaphor.UML.diagram import DiagramCanvas
-from gaphor.core.modeling.event import DiagramItemCreated
 """
 
 # Make getitem behave more politely
